feat(moderation commands): moderation tab made with proper separate folder
in this commit is also, the whole system of the bank and of warns, bans, and everything regarding
moderation. Included is a template for uploading images to an api that will save the images forever.
Also included is the SQLite system to store user information whether warns or some data about it.
The mutation time system was also made with a script that checks mutated users. Ban and Kick's
commands are pulled straight from the api. The Date system is international and adjusted for each
device in its native time.